added pods and migrated bettersegmentcontrol to newest version

+++ Aplikacja Mobilna WZR/Controller/FullTimeControllers/LecturersTimeTableController.swift
@@ -186,8 +186,8 @@ class LecturersTimeTableController: UIViewController {
                 let swipedRight = GesturesUtil.gesturesForFullTImeTablesDecreasing(data: GesturesClass(week: week, day: day, weekForPicker: Int(weekPicker.index), dayForPicker: Int(dayPicker.index)))
                 week = swipedRight.week
                 day = swipedRight.day
-                weekPicker.setIndex(UInt(swipedRight.weekForPicker))
-                dayPicker.setIndex(UInt(swipedRight.dayForPicker))
+                weekPicker.setIndex(swipedRight.weekForPicker)
+                dayPicker.setIndex(swipedRight.dayForPicker)
                 getCurrentLecturersData{}
                 lecturersTableView.reloadData()
                 
@@ -196,8 +196,8 @@ class LecturersTimeTableController: UIViewController {
                 let swipedLeft = GesturesUtil.gesturesForFullTimeTimeTablesAdding(data: GesturesClass(week: week, day: day, weekForPicker: Int(weekPicker.index), dayForPicker: Int(dayPicker.index)))
                 week = swipedLeft.week
                 day = swipedLeft.day
-                weekPicker.setIndex(UInt(swipedLeft.weekForPicker))
-                dayPicker.setIndex(UInt(swipedLeft.dayForPicker))
+                weekPicker.setIndex(swipedLeft.weekForPicker)
+                dayPicker.setIndex(swipedLeft.dayForPicker)
                 getCurrentLecturersData{}
                 lecturersTableView.reloadData()
